backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.1
django-filter==23.4
djangorestframework==3.14.0
flake8==6.1.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
Markdown==3.5.1
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.1.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==2.1.0
zipp==3.17.0
